+++ RAnalogs/RAnEn/src/RcppExports.cpp
@@ -31,8 +31,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // generateAnalogs
-List generateAnalogs(NumericVector R_test_forecasts, NumericVector R_test_forecasts_dims, NumericVector R_test_forecasts_station_x, NumericVector R_test_forecasts_station_y, NumericVector R_test_times, NumericVector R_search_forecasts, NumericVector R_search_forecasts_dims, NumericVector R_search_forecasts_station_x, NumericVector R_search_forecasts_station_y, NumericVector R_search_times, NumericVector R_flts, NumericVector R_search_observations, NumericVector R_search_observations_dims, NumericVector R_observation_times, size_t num_members, size_t observation_id, bool quick, IntegerVector R_circulars, bool search_extension, bool extend_observations, bool preserve_similarity, bool preserve_mapping, bool preserve_search_stations, size_t max_num_search_stations, double distance, size_t num_nearest_stations, int time_match_mode, double max_par_nan, double max_flt_nan, NumericVector R_test_times_compare, NumericVector R_search_times_compare, bool operational, int verbose);
-RcppExport SEXP _RAnEn_generateAnalogs(SEXP R_test_forecastsSEXP, SEXP R_test_forecasts_dimsSEXP, SEXP R_test_forecasts_station_xSEXP, SEXP R_test_forecasts_station_ySEXP, SEXP R_test_timesSEXP, SEXP R_search_forecastsSEXP, SEXP R_search_forecasts_dimsSEXP, SEXP R_search_forecasts_station_xSEXP, SEXP R_search_forecasts_station_ySEXP, SEXP R_search_timesSEXP, SEXP R_fltsSEXP, SEXP R_search_observationsSEXP, SEXP R_search_observations_dimsSEXP, SEXP R_observation_timesSEXP, SEXP num_membersSEXP, SEXP observation_idSEXP, SEXP quickSEXP, SEXP R_circularsSEXP, SEXP search_extensionSEXP, SEXP extend_observationsSEXP, SEXP preserve_similaritySEXP, SEXP preserve_mappingSEXP, SEXP preserve_search_stationsSEXP, SEXP max_num_search_stationsSEXP, SEXP distanceSEXP, SEXP num_nearest_stationsSEXP, SEXP time_match_modeSEXP, SEXP max_par_nanSEXP, SEXP max_flt_nanSEXP, SEXP R_test_times_compareSEXP, SEXP R_search_times_compareSEXP, SEXP operationalSEXP, SEXP verboseSEXP) {
+List generateAnalogs(NumericVector R_test_forecasts, NumericVector R_test_forecasts_dims, NumericVector R_test_forecasts_station_x, NumericVector R_test_forecasts_station_y, NumericVector R_test_times, NumericVector R_search_forecasts, NumericVector R_search_forecasts_dims, NumericVector R_search_forecasts_station_x, NumericVector R_search_forecasts_station_y, NumericVector R_search_times, NumericVector R_flts, NumericVector R_search_observations, NumericVector R_search_observations_dims, NumericVector R_observation_times, NumericVector R_weights, size_t num_members, size_t observation_id, bool quick, IntegerVector R_circulars, bool search_extension, bool extend_observations, bool preserve_similarity, bool preserve_mapping, bool preserve_search_stations, size_t max_num_search_stations, double distance, size_t num_nearest_stations, int time_match_mode, double max_par_nan, double max_flt_nan, NumericVector R_test_times_compare, NumericVector R_search_times_compare, bool operational, int verbose);
+RcppExport SEXP _RAnEn_generateAnalogs(SEXP R_test_forecastsSEXP, SEXP R_test_forecasts_dimsSEXP, SEXP R_test_forecasts_station_xSEXP, SEXP R_test_forecasts_station_ySEXP, SEXP R_test_timesSEXP, SEXP R_search_forecastsSEXP, SEXP R_search_forecasts_dimsSEXP, SEXP R_search_forecasts_station_xSEXP, SEXP R_search_forecasts_station_ySEXP, SEXP R_search_timesSEXP, SEXP R_fltsSEXP, SEXP R_search_observationsSEXP, SEXP R_search_observations_dimsSEXP, SEXP R_observation_timesSEXP, SEXP R_weightsSEXP, SEXP num_membersSEXP, SEXP observation_idSEXP, SEXP quickSEXP, SEXP R_circularsSEXP, SEXP search_extensionSEXP, SEXP extend_observationsSEXP, SEXP preserve_similaritySEXP, SEXP preserve_mappingSEXP, SEXP preserve_search_stationsSEXP, SEXP max_num_search_stationsSEXP, SEXP distanceSEXP, SEXP num_nearest_stationsSEXP, SEXP time_match_modeSEXP, SEXP max_par_nanSEXP, SEXP max_flt_nanSEXP, SEXP R_test_times_compareSEXP, SEXP R_search_times_compareSEXP, SEXP operationalSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -50,6 +50,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericVector >::type R_search_observations(R_search_observationsSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type R_search_observations_dims(R_search_observations_dimsSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type R_observation_times(R_observation_timesSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type R_weights(R_weightsSEXP);
     Rcpp::traits::input_parameter< size_t >::type num_members(num_membersSEXP);
     Rcpp::traits::input_parameter< size_t >::type observation_id(observation_idSEXP);
     Rcpp::traits::input_parameter< bool >::type quick(quickSEXP);
@@ -69,7 +70,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericVector >::type R_search_times_compare(R_search_times_compareSEXP);
     Rcpp::traits::input_parameter< bool >::type operational(operationalSEXP);
     Rcpp::traits::input_parameter< int >::type verbose(verboseSEXP);
-    rcpp_result_gen = Rcpp::wrap(generateAnalogs(R_test_forecasts, R_test_forecasts_dims, R_test_forecasts_station_x, R_test_forecasts_station_y, R_test_times, R_search_forecasts, R_search_forecasts_dims, R_search_forecasts_station_x, R_search_forecasts_station_y, R_search_times, R_flts, R_search_observations, R_search_observations_dims, R_observation_times, num_members, observation_id, quick, R_circulars, search_extension, extend_observations, preserve_similarity, preserve_mapping, preserve_search_stations, max_num_search_stations, distance, num_nearest_stations, time_match_mode, max_par_nan, max_flt_nan, R_test_times_compare, R_search_times_compare, operational, verbose));
+    rcpp_result_gen = Rcpp::wrap(generateAnalogs(R_test_forecasts, R_test_forecasts_dims, R_test_forecasts_station_x, R_test_forecasts_station_y, R_test_times, R_search_forecasts, R_search_forecasts_dims, R_search_forecasts_station_x, R_search_forecasts_station_y, R_search_times, R_flts, R_search_observations, R_search_observations_dims, R_observation_times, R_weights, num_members, observation_id, quick, R_circulars, search_extension, extend_observations, preserve_similarity, preserve_mapping, preserve_search_stations, max_num_search_stations, distance, num_nearest_stations, time_match_mode, max_par_nan, max_flt_nan, R_test_times_compare, R_search_times_compare, operational, verbose));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -77,7 +78,7 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_RAnEn_checkOpenMP", (DL_FUNC) &_RAnEn_checkOpenMP, 0},
     {"_RAnEn_generateTimeMapping", (DL_FUNC) &_RAnEn_generateTimeMapping, 5},
-    {"_RAnEn_generateAnalogs", (DL_FUNC) &_RAnEn_generateAnalogs, 33},
+    {"_RAnEn_generateAnalogs", (DL_FUNC) &_RAnEn_generateAnalogs, 34},
     {NULL, NULL, 0}
 };